Override all printable single-byte chars
This code was copied from isearch.el. Wider-range characters are less
likely to be bound by modes, and those should be covered by the
rebinding of self-insert.

This means that god-mode is now usable with magit-mode, dired-mode,
etc. without issue. I've left the exempt modes functionality as-is. It's
still convenient to not activate god-mode by default for some buffers,
like dired-mode or magit mode. But it means unlike in the past where by
habbit you would activate god-mode on these kinds of
buffers (e.g. magit, magit's commit viewer, dired) and end up running
the special-mode's commands insteaad of what you wanted.

When updating to this, remember to re-apply any keybindings you have
locally for god-local-map.

Pinging everyone who I know to be using god-mode, as this makes for a
substantial improvement.

@fgallina @magnars @dillonkearns @pyluyten @YangZhao11 @ruediger
@bradleywright @conorbev @dragonwasrobot @ArneBab @fgeller @dennisgroves
@tobytripp @ljos @syohex @priyadarshan @rprimus @stardiviner @darth10
@j0xaf @rlazo @tuhdo @Crazy @theoreticMK @dhaley @AtticHacker @huchunyu
@eerie