Update for upcoming Magit 2.1.0 release
Modify exemption rules to work with the upcoming Magit 2.1.0, which is
scheduled to be released on July 1 [1].

The main modes that need to be accounted for are magit-pop-mode,
magit-popup-sequence-mode, and git-commit-mode.  Use a predicate
function for the check because git-commit-mode is no longer defined as a
major mode.

The new function also checks for children of magit-mode, but this is not
actually required because magit-mode is derived from special-mode, so
all the remaining Magit modes already are covered by god-special-mode-p.

These changes should still work with the previous version of Magit.

[1]: https://github.com/magit/magit/issues/1645