Fix god-mode-upper-p's use of god-mode-alist
The relatively recent change to remove hard-coding the `?G` character broke handling of that character. It impacted me because if I set mark, then moved point forward by a word, and then moved point forward by a sexp using `G f`, mark would be reset to the location of point when I issued the `G f` command.

For example, say you have the string "Here we go" in a buffer. Put point before the word "Here", then, in `god-mode`, hit space to activate mark, then hit "g f" to move point forward by a word. The region will cover the word "Here". If you then hit "G f", the region will move forward but no longer include include the word "Here".

Looking at the code, there is a problem because the keys of `god-mode-alist` are strings while `key` in `god-mode-upper-p` is a character. We could either make the change I have made here, or make the keys of `god-mode-alist` characters.