Make (god-mode) always work
It's often been the case that I'll hit the `god-mode` key and it does nothing, because `god-global-mode` and `god-local-mode` is out of sync.

Either I wanna turn on god-mode, because it's off locally, but it's on globally (and there's no indication to show that), so the key (seemingly) does nothing. It only affects background buffers (and sometimes in ways I don't even want).

Or, I wanna turn off god-mode, but unbeknownst to me it's already off globally, it's just off locally. So again the key does nothing in the active buffer, only weirds up the background buffers.

With this patch, if `god-global-mode` and `god-local-mode` has gotten out of sync, this key syncs them up by changing the local-buffer which is always what I want. If they are already in sync, it toggles them (i.e. the old behavior, which is exactly what you want when they are in sync).