Added evil-god-toggle to list of options for evil (#162)
`evil-god-toggle` (https://github.com/jam1015/evil-god-toggle)
is a package I wrote that expands on evil-god-state.
Added it to the list of ways to use god-mode with evil.
Offers features of evil-god state plus better transferring
of visual mode to active regions, toggling across all buffers,
god-off state which is like emacs-state but easier to switch
in and out of.